backend/pyproject.toml
[project]
name = "langgraph-chatkit-integration"
version = "0.1.0"
description = "LangGraph ChatKit Integration - ChatKit backend powered by LangGraph API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "httpx>=0.28,<0.29",
    "uvicorn[standard]>=0.36,<0.37",
    "openai-chatkit>=1.0.2,<2",
    "pydantic>=2.0,<3",
    "itsdangerous>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "chatkit_langgraph*", "examples*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
